refactor!: move liftover to separate module
close #158

* Remove LiftOver from UtaDatabase class
* LiftOver separated into its own class
* ManeTranscript and ExonGenomicCoordsMapper now take liftover as param in init method
