+++ src/cool_seq_tool/mappers/__init__.py
@@ -1,8 +1,9 @@
 """Module for mapping data"""
 
 from .alignment import AlignmentMapper  # noqa: I001
+from .liftover import LiftOver
 from .mane_transcript import ManeTranscript
 from .exon_genomic_coords import ExonGenomicCoordsMapper
 
 
-__all__ = ["AlignmentMapper", "ManeTranscript", "ExonGenomicCoordsMapper"]
+__all__ = ["AlignmentMapper", "LiftOver", "ManeTranscript", "ExonGenomicCoordsMapper"]
